[SPARK-39731][SQL] Fix issue in CSV and JSON data sources when parsing dates in "yyyyMMdd" format with CORRECTED time parser policy
### What changes were proposed in this pull request?

This PR fixes a correctness issue when reading a CSV or a JSON file with dates in "yyyyMMdd" format:
```
name,mydate
1,2020011
2,20201203
```
or
```
{"date": "2020011"}
{"date": "20201203"}
```

Prior to https://github.com/apache/spark/pull/32959, reading this CSV file would return:
```
+----+--------------+
|name|mydate        |
+----+--------------+
|1   |null          |
|2   |2020-12-03    |
+----+--------------+
```

However, after the patch, the invalid date is parsed because of the much more lenient parsing in `DateTimeUtils.stringToDate`, the method treats `2020011` as a full year:
```
+----+--------------+
|name|mydate        |
+----+--------------+
|1   |+2020011-01-01|
|2   |2020-12-03    |
+----+--------------+
```
Similar result would be observed in JSON.

This PR attempts to address correctness issue by introducing a new configuration option `enableDateTimeParsingFallback` which allows to enable/disable the backward compatible parsing.

Currently, by default we will fall back to the backward compatible behavior only if parser policy is legacy and no custom pattern was set (this is defined in `UnivocityParser` and `JacksonParser` for csv and json respectively).

### Why are the changes needed?
Fixes a correctness issue in Spark 3.4.

### Does this PR introduce _any_ user-facing change?

In order to avoid correctness issues when reading CSV or JSON files with a custom pattern, a new configuration option `enableDateTimeParsingFallback` has been added to control whether or not the code would fall back to the backward compatible behavior of parsing dates and timestamps in CSV and JSON data sources.
- If the config is enabled and the date cannot be parsed, we will fall back to `DateTimeUtils.stringToDate`.
- If the config is enabled and the timestamp cannot be parsed, `DateTimeUtils.stringToTimestamp` will be used.
- Otherwise, depending on the parser policy and a custom pattern, the value will be parsed as null.

### How was this patch tested?

I added unit tests for CSV and JSON to verify the fix and the config option.

Closes #37147 from sadikovi/fix-csv-date-inference.

Authored-by: Ivan Sadikov <ivan.sadikov@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>