[SPARK-37194][SQL] Avoid unnecessary sort in v1 write if it's not dynamic partition
### What changes were proposed in this pull request?

This is a rework for https://github.com/apache/spark/pull/34468, since we pull out v1write required ordering.

This prs add a new parameter numStaticPartitions to v1write and FileFormatWriter so we can skip unnecessary local sort for static partition write.

### Why are the changes needed?

The v1 write requires ordering for dynamic partition, bucket expression and sort column during writing. The reason is the `DynamicPartitionDataSingleWriter` and `DynamicPartitionDataConcurrentWriter` assume the partition and bucket columns are continuous. Then if partition column is static, it's unnecessary to do the local sort.

For  v1 write, `InsertIntoHadoopFsRelationCommand` is the only case which adds a local sort even if the partition column is static.

### Does this PR introduce _any_ user-facing change?

no, only improve performance

### How was this patch tested?

add test

Closes #37290 from ulysses-you/v1write.

Authored-by: ulysses-you <ulyssesyou18@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>