[SPARK-39902][SQL] Add Scan details to spark plan scan node in SparkUI
### What changes were proposed in this pull request?

In this PR, we propose to add a method "String name()" to the Scan interface, that "BatchScanExec" can invoke to set the node name of the plan. This nodeName will be eventually used by "SparkPlanGraphNode" to display it in the header of the UI node.

### Why are the changes needed?

Since for DSv2, the scan node in the spark plan on SparkUI simply shows "BatchScan" instead of an informative name.

### Does this PR introduce _any_ user-facing change?

Yes, after this change the user will be able to see the scan name in the spark plan on SparkUI.

### How was this patch tested?

- Tested this change using existing UTs
- Further tested this change by overriding the newly added method in Iceberg's implementation of the Scan interface.

Closes #37325 from sumeetgajjar/v2_scan_ui_improvement.

Authored-by: Sumeet Gajjar <sumeetgajjar93@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>