[SPARK-39940][SS] Refresh catalog table on streaming query with DSv1 sink
Credit to pranavanand on figuring out the issue and providing the broken test code!

### What changes were proposed in this pull request?

This PR proposes to refresh the destination catalog table when streaming query is writing to the catalog table via DSv1 sink.

### Why are the changes needed?

It has been a long standing issue that streaming query is not aware of catalog table (not the table brought by DSv2), hence no way for streaming query to refresh the catalog table after updating it. As a side effect of SPARK-39564, Spark brings up the information of destination catalog table into MicrobatchExecution, which enables to refresh table.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

New UT.

Closes #37368 from HeartSaVioR/SPARK-39940.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>