[SPARK-39900][SQL] Address partial or negated condition in binary format's predicate pushdown
### What changes were proposed in this pull request?

fix `BinaryFileFormat` filter push down bug.

Before modification, when Filter tree is:
````
-Not
- - IsNotNull
````

Since `IsNotNull` cannot be matched, `IsNotNull` will return a result that is always true (that is, `case _ => (_ => true)`), that is, no filter pushdown is performed. But because there is still a `Not`, after negation, it will return a result that is always False, that is, no result can be returned.

### Why are the changes needed?

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

test suit in `BinaryFileFormatSuite`

```
    testCreateFilterFunction(
      Seq(Not(IsNull(LENGTH))),
      Seq((t1, true), (t2, true), (t3, true)))
```

Closes #37350 from zzzzming95/SPARK-39900.

Lead-authored-by: zzzzming95 <505306252@qq.com>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>