[MINOR][SQL] Improve the comments about null tracking for UnsafeRow
### What changes were proposed in this pull request?
This PR expect to improve the comments about null tracking for `UnsafeRow`.
The old comment of `UnsafeRow` have confused text `[null bit set]`.
In fact, the portion is a lot of bit or bit array which does't always be null.
On the other hand, it tell users nothing. we need the information more clear.

### Why are the changes needed?
Improve the comments about null tracking for `UnsafeRow`.

### Does this PR introduce _any_ user-facing change?
'No'.
Just update comments.

### How was this patch tested?
N/A

Closes #37340 from beliefer/UnsafeRow_comment.

Authored-by: Jiaan Geng <beliefer@163.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>