[SPARK-40006][PYTHON][DOCS][FOLLOW-UP] Remove unused Spark context and duplicated Spark session initialization
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/37437 which missed to remove unused `sc` and duplicated Spark session initialization.

### Why are the changes needed?

To make the consistent example, and remove unused variables.

### Does this PR introduce _any_ user-facing change?

No. It's a documentation change. However, the previous PR is not released yet.

### How was this patch tested?

Ci in this PR should test it out.

Closes #37457 from HyukjinKwon/SPARK-40006-followup.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>