[SPARK-40010][PYTHON][DOCS] Make pyspark.sql.window examples self-contained
### What changes were proposed in this pull request?

This PR proposes to improve the examples in `pyspark.sql.window` by making each example self-contained with a brief explanation and a bit more realistic example.

### Why are the changes needed?

To make the documentation more readable and able to copy and paste directly in PySpark shell.

### Does this PR introduce _any_ user-facing change?

Yes, it changes the documentation

### How was this patch tested?

Manually ran each doctest.

Closes #37450 from dcoliversun/SPARK-40010.

Authored-by: Qian.Sun <qian.sun2020@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>