[SPARK-40058][CORE] Avoid filter file path more than once in HadoopFSUtils
### What changes were proposed in this pull request?
Refactor path filter logic in HadoopFSUtils to avoid the same filter logic is applied to a file multiple time. Method listLeafFiles is called recursively. Especially, this filter will be used in single thread on all files at driver side. This will lead to a performance issue when the filter logic is heavy.

### Why are the changes needed?
Apply filter only on filestatus as soon as they are firstly met.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
No test was added as such change is simple enough.

Closes #37498 from guanziyue/SPARK-40058.

Authored-by: guanziyue <30882822+guanziyue@users.noreply.github.com>
Signed-off-by: Sean Owen <srowen@gmail.com>