[SPARK-37980][SQL] Extend METADATA column to support row indexes for Parquet files
### What changes were proposed in this pull request?

This change adds `row_index` column to `_metadata` struct. This column allows us to uniquely identify rows read from a given file with an index number. The n-th row  in a given file with be assigned `n-1` row index in every scan of the file, irrespective of file splitting and data skipping in use.

The new column requires file format specific support. This change introduces Parquet support, and other formats can follow later.

### Why are the changes needed?

Row Indexes can be used in a variety of ways. A (fileName, rowIndex) tuple uniquely identifies a row in a table. This information can be used to mark rows e.g. can be used to create an indexer.

### Does this PR introduce _any_ user-facing change?

Yes. With this change the customers will be able to access `_metadata.row_index` metadata column when reading Parquet data. The schema of `_matadata` column remains unchanged for file formats without row index support.

### How was this patch tested?

* Added `FileMetadataStructSuite.scala` to make sure the feature works correctly in different scenarios (supported/unsupported file format, batch/record reads, on/off heap memory...).
* Added `ParquetRowIndexSuite.scala` to make sure the row indexes are generated correctly for Parquet file in conjunction with any combination of data skipping features.
* Extended `FileMetadataStructRowIndexSuite` to account for new column in `_metadata` struct.

Closes #37228 from ala/row-idx-v4.

Lead-authored-by: Ala Luszczak <ala@databricks.com>
Co-authored-by: IonutBoicu <ionut.boicu@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>