[SPARK-40136][SQL] Fix the fragment of SQL query contexts
### What changes were proposed in this pull request?
In the PR, I propose to change implementation of `fragment` in `SQLQueryContext` to get the fragment correctly. Before the changes, `fragment` misses the last char because `originStopIndex.get` points to the last char of fragments but `substring()` gets a sub-string up to `endIndex - 1`.

### Why are the changes needed?
It fixes a bug.

### Does this PR introduce _any_ user-facing change?
Yes.

### How was this patch tested?
By running new test suite:
```
$ build/sbt "test:testOnly *SQLQueryContextSuite"
$ build/sbt "test:testOnly *TreeNodeSuite"
$ build/sbt "test:testOnly *QueryExecutionAnsiErrorsSuite"
```

Closes #37566 from MaxGekk/fix-fragment.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>