[SPARK-40120][PYTHON][SQL][DOCS] Make pyspark.sql.readwriter examples self-contained
### What changes were proposed in this pull request?

This PR proposes to improve the examples in `pyspark.sql.readwriter` by making each example self-contained with a brief explanation and a bit more realistic example.

### Why are the changes needed?

To make the documentation more readable and able to copy and paste directly in PySpark shell.

### Does this PR introduce _any_ user-facing change?

Yes, it changes the documentation

### How was this patch tested?

Manually ran each doctests. CI also runs this.

Closes #37570 from HyukjinKwon/SPARK-40120.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>