[SPARK-40089][SQL] Fix sorting for some Decimal types
### What changes were proposed in this pull request?
This fixes https://issues.apache.org/jira/browse/SPARK-40089 where the prefix can overflow in some cases and the code assumes that the overflow is always on the negative side, not the positive side.

### Why are the changes needed?
This adds a check when the overflow does happen to know what is the proper prefix to return.

### Does this PR introduce _any_ user-facing change?
No, unless you consider getting the sort order correct a user facing change.

### How was this patch tested?
I tested manually with the file in the JIRA and I added a small unit test.

Closes #37540 from revans2/fix_dec_sort.

Authored-by: Robert (Bobby) Evans <bobby@apache.org>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>