[SPARK-40191][PYTHON][CORE][DOCS] Make pyspark.resource examples self-contained
### What changes were proposed in this pull request?

This PR proposes to add a working example `pyspark.resource.ResourceProfile`

In addition, this PR adds return and parameter descriptions with fixing a typo in Scaladoc side.

### Why are the changes needed?

To make the documentation more readable and able to copy and paste directly in PySpark shell.

### Does this PR introduce _any_ user-facing change?

Yes, it changes the documentation

### How was this patch tested?

Manually ran each doctests. CI also runs this.

Closes #37627 from HyukjinKwon/SPARK-40191.

Lead-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Co-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>