[SPARK-40220][SQL] Don't output the empty map of error message parameters
### What changes were proposed in this pull request?
In the PR, I propose to output the JSON map `messageParameters` of error messages in the MINIMAL and STANDARD formats only when the map is not empty.

### Why are the changes needed?
To be consistent w/ output of other JSON fields of error messages.

### Does this PR introduce _any_ user-facing change?
Yes. The PR can change user-facing error message in the MINIMAL and STANDARD formats of error messages.

### How was this patch tested?
By running the affected test suites:
```
$ build/sbt "sql/testOnly org.apache.spark.sql.SQLQueryTestSuite"
```

Closes #37660 from MaxGekk/no-empty-messageParameters.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>