[SPARK-40160][PYTHON][DOCS] Make pyspark.broadcast examples self-contained
### What changes were proposed in this pull request?

This PR proposes to improve the examples in `pyspark.broadcast` by making each example self-contained with a brief explanation and a bit more realistic example.

### Why are the changes needed?

To make the documentation more readable and able to copy and paste directly in PySpark shell.

### Does this PR introduce _any_ user-facing change?

Yes, it changes the documentation.

### How was this patch tested?

Manually ran each doctest.

Closes #37629 from dcoliversun/SPARK-40160.

Authored-by: Qian.Sun <qian.sun2020@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>