[SPARK-40315][SQL] Add hashCode() for Literal of ArrayBasedMapData
### What changes were proposed in this pull request?
There is no explicit `hashCode()` function override for `ArrayBasedMapData`. As a result, there is a non-deterministic error where the `hashCode()` computed for `Literal`s of `ArrayBasedMapData` can be different for two equal objects (`Literal`s of `ArrayBasedMapData` with equal keys and values).

In this PR, we add a `hashCode` function so that it works exactly as we expect.

### Why are the changes needed?
This is a bug fix for a non-deterministic error. It is also more consistent with the rest of Spark if we implement the `hashCode` method instead of relying on defaults. We can't add the `hashCode` directly to `ArrayBasedMapData` because of SPARK-9415.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
A simple unit test was added.

Closes #37807 from c27kwan/SPARK-40315-lit.

Authored-by: Carmen Kwan <carmen.kwan@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>