[SPARK-40293][SQL] Make the V2 table error message more meaningful
### What changes were proposed in this pull request?
Add more info in the v2 table error message to make it more meaningful.

### Why are the changes needed?
When V2 catalog is not configured, Spark fails to access/create a table using the V2 API and silently falls back to attempting to do the same operation using the V1 Api. This happens frequently among the users. We want to have a better error message so that users can fix the configuration/usage issue by themselves.

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
existing tests

Closes #37746 from huaxingao/error_message.

Authored-by: huaxingao <huaxin_gao@apple.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>