[SPARK-40411][SS] Refactor FlatMapGroupsWithStateExec to have a parent trait
### What changes were proposed in this pull request?

This PR proposes to factor the common attributes out from `FlatMapGroupsWithStateExec` to `FlatMapGroupsWithStateExecBase`.

### Why are the changes needed?

There are a lot of stuff to share if you implement another version of `FlatMapGroupsWithStateExec`.
Should better factor them out. This is also part of https://github.com/apache/spark/pull/37285 which demonstrates how the refactored trait is used.

### Does this PR introduce _any_ user-facing change?

No, this is refactoring-only.

### How was this patch tested?

Existing test cases should cover it.

Closes #37859 from HyukjinKwon/SPARK-40411.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>