[SPARK-40466][SS] Improve the error message when DSv2 is disabled while DSv1 is not avaliable
…le DSv1 is not avaliable.

### What changes were proposed in this pull request?

Improve the error message when DSv2 is disable while its fallback DSv1 is not available.

### Why are the changes needed?

Improve the user experience. When users get UnsupportOperationError for the disabled DSv2, they are able to know which config to modify to enable the V2 source.

### Does this PR introduce _any_ user-facing change?

Yes, error message.

### How was this patch tested?

N/A, just the message change

Closes #37917 from huanliwang-db/SPARK-40466.

Authored-by: Huanli Wang <huanli.wang@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>