[SPARK-40580][PS][DOCS] Update the document for `DataFrame.to_orc`
### What changes were proposed in this pull request?

This PR proposes to update the docstring of `DataFrame.to_orc`, since `pandas.DataFrame.to_orc` is supported from pandas 1.5.0, but the behavior is a bit different.

### Why are the changes needed?

From pandas 1.5.0, they support [writing the DataFrame to ORC files](https://pandas.pydata.org/docs/whatsnew/v1.5.0.html#writing-to-orc-files).

In pandas API on Spark, we already support this feature, but the behavior is different.

So, we should mention the difference in the documents.

### Does this PR introduce _any_ user-facing change?

Yes, documentation update.

### How was this patch tested?

The existing doctest should pass.

Closes #38018 from itholic/SPARK-40580.

Lead-authored-by: itholic <haejoon.lee@databricks.com>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>