[MINOR][CONNECT][PYTHON] Miscellaneous cleanup in Spark Connect componenet
### What changes were proposed in this pull request?

This PR proposes:
1. Fix the code style in `SparkConnectPlugin.scala` and `SparkBuild.scala` to be consistent with others.
2. Rename `data_frame` to `dataframe` to be consistent with existing PySpark codes.

This Pr is a sort of a followup of https://github.com/apache/spark/pull/37710

### Why are the changes needed?

To follow existing codebase, and style.

### Does this PR introduce _any_ user-facing change?

No, the codes are not released yet. The only notable change would be renaming `data_frame` to `dataframe` to be consistent.

### How was this patch tested?

Ci in this PR should validate the changes.

Closes #38121 from HyukjinKwon/minor-cleanup.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>