[SPARK-40682][SQL][TESTS] Set `spark.driver.maxResultSize` to 3g in `SqlBasedBenchmark`
### What changes were proposed in this pull request?

This PR aims to set `spark.driver.maxResultSize` to `3g` from the default value, `1g`,  in `SqlBasedBenchmark`.

### Why are the changes needed?

Apache Spark benchmark is using `4g` JVM memory. We can utilize it for `spark.driver.maxResultSize` in some cases after this PR.

### Does this PR introduce _any_ user-facing change?

No. This is a test case and unblocks only the cases which failed before.

### How was this patch tested?

N/A

Closes #38136 from dongjoon-hyun/SPARK-40682.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>