[SPARK-40454][CONNECT] Initial DSL framework for protobuf testing
### What changes were proposed in this pull request?

Implement an approach to testing the proto to Scala conversion with a DSL according to the proposal in [spark-connect-testing-plan](https://docs.google.com/document/d/1n6EgS5vcmbwJUs5KGX4PzjKZVcSKd0qf0gLNZ6NFvOE/edit#bookmark=id.v673sn6dm9k4).

More specifically, we need to decouple the testing for clients' data frame API and for the proto contract. The testing for proto is used to protect that any core changes in SQL does not break proto contracts. If necessary, a change in SQL plans has to justify that it should work with connect proto but does not need to cover clients (given that there could be many clients).  This DSL proposal is a part of that effort.

This is still experimental and could face changes in the future.

### Why are the changes needed?

For testing the connect protobuf contract.

### Does this PR introduce _any_ user-facing change?

NO

### How was this patch tested?

UT

Closes #37994 from amaliujia/SPARK-40454.

Authored-by: Rui Wang <rui.wang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>