[SPARK-40448][CONNECT][FOLLOWUP] Use more suitable variable name and fix code style
### What changes were proposed in this pull request?
We are very concerned about the progress of the Spark Connect, so we understand the current draft and read preliminary implementation. We can see that some code style are inappropriate, or even cause confusion. Therefore, this PR try to correct it.

### Why are the changes needed?
Fix code style on scala server side.

### Does this PR introduce _any_ user-facing change?
'No'.
Spark Connect just started.

### How was this patch tested?
N/A

Closes #38061 from beliefer/SPARK-40448_followup.

Authored-by: Jiaan Geng <beliefer@163.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>