[SPARK-40799][BUILD][CONNECT] Enforce scalafmt for Spark Connect module
### What changes were proposed in this pull request?
This changes enables enforcing `scalafmt` for the Spark Connect module to have a consistent enforced coding style and reduce unnecessary code-review cycles. This patch specifically only enforces this on the Spark Connect module to avoid impact on the other areas of the Spark codebase.

To pass the pre-commit, this change applies `scalafmt` on the existing code-base.

### Why are the changes needed?
Faster, focussed code reviews.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Manual testing / CI

Closes #38258 from grundprinzip/spark-40799.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>