[SPARK-39783][SQL] Quote qualifiedName to fix backticks for column candidates in error messages
### What changes were proposed in this pull request?

The `NamedExpression.qualifiedName` is a concatenation of qualifiers and the name, joined by `dots`. If those contain `dots`, the result `qualifiedName` is ambiguous. Quoting those if they contain `dots` fixes this, while this also fixes quoting column candidates in the error messages `UNRESOLVED_COLUMN` and `UNRESOLVED_MAP_KEY`:

`UNRESOLVED_COLUMN`:
```
Seq((0)).toDF("the.id").select("the.id").show()
```

The error message should read

    org.apache.spark.sql.AnalysisException: [UNRESOLVED_COLUMN] A column or function parameter
    with name `the`.`id` cannot be resolved. Did you mean one of the following? [`the.id`];

while it was:

    org.apache.spark.sql.AnalysisException: [UNRESOLVED_COLUMN] A column or function parameter
    with name `the`.`id` cannot be resolved. Did you mean one of the following? [`the`.`id`];

`UNRESOLVED_MAP_KEY`:
```
Seq((0)).toDF("id")
  .select(map(lit("key"), lit(1)).as("map"), lit(2).as("other.column"))
  .select($"`map`"($"nonexisting")).show()
```

The error message should read

    Cannot resolve column `nonexisting` as a map key. If the key is a string literal, please add single quotes around it.
    Otherwise did you mean one of the following column(s)? [`map`, `other.column`];

while it was:

    Cannot resolve column `nonexisting` as a map key. If the key is a string literal, please add single quotes around it.
    Otherwise did you mean one of the following column(s)? [`map`, `other`.`column`];

### Why are the changes needed?
The current quoting is wrong and `qualifiedName` is ambiguous if `name` or `qualifiers` contain `dots`.

### Does this PR introduce _any_ user-facing change?
It corrects the error message.

### How was this patch tested?
This is tested in `AnalysisErrorSuite`, `DatasetSuite` and `QueryCompilationErrorsSuite.scala`.

Closes #38256 from EnricoMi/branch-correct-backticks-error-message.

Authored-by: Enrico Minack <github@enrico.minack.dev>
Signed-off-by: Max Gekk <max.gekk@gmail.com>